random: move XoRoShiRo128PlusPlus into random module
This is preparation for making it more generally accessible.
