+++ src/test/fuzz/bip324.cpp
@@ -4,11 +4,11 @@
 
 #include <bip324.h>
 #include <chainparams.h>
+#include <random.h>
 #include <span.h>
 #include <test/fuzz/FuzzedDataProvider.h>
 #include <test/fuzz/fuzz.h>
 #include <test/fuzz/util.h>
-#include <test/util/xoroshiro128plusplus.h>
 
 #include <cstdint>
 #include <vector>
