+++ test/sanitizer_suppressions/ubsan
@@ -57,7 +57,8 @@ unsigned-integer-overflow:CBlockPolicyEstimator::processBlockTx
 unsigned-integer-overflow:TxConfirmStats::EstimateMedianVal
 unsigned-integer-overflow:prevector.h
 unsigned-integer-overflow:EvalScript
-unsigned-integer-overflow:xoroshiro128plusplus.h
+unsigned-integer-overflow:XoRoShiRo128PlusPlus::operator()
+unsigned-integer-overflow:XoRoShiRo128PlusPlus::SplitMix64
 unsigned-integer-overflow:bitset_detail::PopCount
 implicit-integer-sign-change:CBlockPolicyEstimator::processBlockTx
 implicit-integer-sign-change:SetStdinEcho
@@ -75,6 +76,6 @@ shift-base:arith_uint256.cpp
 shift-base:crypto/
 shift-base:streams.h
 shift-base:FormatHDKeypath
-shift-base:xoroshiro128plusplus.h
+shift-base:XoRoShiRo128PlusPlus::operator()
 shift-base:RandomMixin<*>::randbits
 shift-base:RandomMixin<*>::randbits<*>
